Rename RangeSpecSeq to ChunkRangesSeq
RangeSpec and RangeSpecSeq are really just optimized representations of
ChunkRanges/sequences of ChunkRanges for the wire protocol. So this is now
relegated to the ser/de impl. In user facing code we always work with
ChunkRanges now.

We might newtype ChunkRanges later to hide some of the weirdness of
range_collections::RangeSet.